Moving maps helpers inside map package
